% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.light.r
\name{calc.light}
\alias{calc.light}
\title{Calculate Position-based Likelihood}
\usage{
calc.light(light = NULL, locs.grid, dateVec, res = 1)
}
\arguments{
\item{light}{is -LightLoc file output from DAP/Tag Portal for WC tags and 
contains tag-measured dawn/dusk times.}

\item{locs.grid}{is list output from \code{setup.locs.grid}}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}

\item{res}{is resolution of light grid in degrees. default is 1 deg. higher resolution (e.g. res = .25 for 1/4 deg) takes considerably longer to compute}
}
\value{
L is a raster of dim(lon x lat x dateVec) containing likelihood surfaces for each
  time point
}
\description{
\code{calc.locs} calculates likelihood estimates for each day's estimated dawn dusk times from the tag
}
\details{
Tag-measured sunrise/sunset (SRSS) times are first filtered according to the min/max times possible. possible values are computed based on spatial limits included in locs.grid input. after filtering, each yearday has a grid of georeferenced SRSS times that the tag-measured times are compared to in order to generate a likelihood.
}

