% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.profile.r
\name{calc.profile}
\alias{calc.profile}
\title{Calculate Depth-temperature profile based likelihood}
\usage{
calc.profile(pdt, dat = NULL, lat = NULL, lon = NULL, dateVec,
  envType = "woa", hycom.dir = NULL)
}
\arguments{
\item{pdt}{is -PDT data from WC psat tag summarizing depth/temperature data
over a programmed time interval}

\item{dat}{is monthly global 1/4deg climatology data from WOA13. NULL if envType = 'hycom'.}

\item{lat}{is vector of latitudes from dat. NULL if envType = 'hycom'.}

\item{lon}{is vector of longitudes from dat. NULL if envType = 'hycom'.}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}

\item{envType}{is character indicating whether to compare tag-based profile to World Ocean Atlas ('woa') or HYCOM ('hycom').}

\item{hycom.dir}{is path to stored HYCOM model outputs if envType = 'hycom'}
}
\value{
raster of likelihoods for depth-temp profile
       matching between tag data and oceanographic profiles
}
\description{
\code{calc.profile} calculates likelihood of animal position based on 
summarized depth-temperature profiles
}
\details{
Tag-based depth-temperature profile summaries are compared to climatological 
profiles from the World Ocean Atlas (WOA) or HYbrid Coordinate Ocean Model (HYCOM) 
and "matched" to generate position likelihoods. This essentially attempts to estimate animal position based on 
the water mass it is in, particularly if extensive diving performs thorough 
sampling of the environment. However, remember the in situ data is being 
compared to climatological means or the results of an oceanographic model.
}

