% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.sst.r
\name{calc.sst}
\alias{calc.sst}
\title{Calculate SST-based likelihood}
\usage{
calc.sst(tag.sst, sst.dir, dateVec)
}
\arguments{
\item{tag.sst}{variable containing tag-collected SST data}

\item{sst.dir}{local directory where remote sensing SST downloads are stored}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}
}
\value{
likelihood is raster brick of likelihood surfaces representing matches
  between tag-based sst and remotely sensed sst maps
}
\description{
\code{calc.sst} compares tag SST to remotely sensed SST and calculates
likelihoods
}
\examples{
none
}
\seealso{
\code{\link{calc.pdt.int}}, \code{\link{calc.ohc}}
}

