% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.pdt.r
\name{extract.pdt}
\alias{extract.pdt}
\title{Extract PDT from Wildlife Computers tag data}
\usage{
extract.pdt(data)
}
\arguments{
\item{data}{data frame read from .csv output of Wildlife Computers DAP processor. File ends in "-PDTs.csv"}
}
\value{
data frame formatted for pdt data
}
\description{
Extract PDT from Wildlife Computers tag data
}
\examples{
none
}

