% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDateFormat.r
\name{findDateFormat}
\alias{findDateFormat}
\title{Determine date format of vector}
\usage{
findDateFormat(dateVec)
}
\arguments{
\item{dateVec}{a vector of dates as in those from -Histos or  -PDTs from WC tags}
}
\value{
dateformat is character string used as input to strptime(format = dateformat)
}
\description{
\code{findDateFormat} determines the date format of a given vector of dates
}
\examples{
none

}

