% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.env.r
\name{get.env}
\alias{get.env}
\title{Download and Read Oceanographic Data}
\usage{
get.env(uniqueDates, ptt, type = NA, spatLim, save.dir = getwd())
}
\arguments{
\item{uniqueDates}{is a POSIXct vector of desired dates}

\item{type}{is a character string indicating whether you're after sea surface temperature 'sst' or hybrid coordinate ocean model 'ohc' data}

\item{spatLim}{is a list of spatial limits as \code{list(xmin, xmax, ymin, ymax)}}

\item{save.dir}{is the directory to save the downloaded data to}
}
\value{
nothing, just downloads the data to your local machine
}
\description{
\code{get.env} accesses oceanographic data like sea surface temperature from a remote server and downloads the temporal and spatial extent of interest for further use
}
\examples{
splim <- list(xmin=-60, xmax=-40, ymin=10, ymax=30)
get.env('2015-01-01', type = 'sst', splim, getwd())
}

