% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.hycom.R
\name{get.hycom}
\alias{get.hycom}
\title{Download HYCOM data}
\usage{
get.hycom(limits, time, vars = c("water_temp"), include_latlon = TRUE,
  filename = "", type = "r", download.file = TRUE, dir = getwd())
}
\arguments{
\item{limits}{A list of length 4; minlon, maxlon, minlat, maxlat. Longitude values are -180,180}

\item{time}{A vector of length 2 with the minimum and maximum times in form 
\code{as.Date(date)}.}

\item{vars}{A list of variables to download. This should only contain 
'surf_el', 'salinity', 'water_u', water_v','water_temp' but is not checked 
for errors. We currently only use 'water_temp'.}

\item{include_latlon}{Should the array of latitude and longitude values be 
included?}

\item{filename}{An optional filename. If provided, then the data is 
downloaded to that file. Otherwise the data is not downloaded and the url 
is returned.}

\item{type}{indicates type of HYCOM product to download. 'r' is reanalysis 
and 'a' is analysis. see \url{https://hycom.org/dataserver/} for details.}

\item{download.file}{Logical. Should use the default \code{download.file} to 
query the server and download or use the optional \code{curl}. Some users 
may need to use \code{curl} in order to get this to work.}

\item{dir}{is local directory where ncdf files should be downloaded to. 
default is current working directory. if enter a directory that doesn't 
exist, it will be created.}
}
\value{
The url used to extract the requested data from the NetCDF subset 
  service.
}
\description{
\code{get.hycom} downloads HYbrid Coordinate Ocean Model (HYCOM) data for 
given temporal and spatial constraints of your data.
}
\details{
The method may return before the download is completed. It will continue to 
display progress  until the download completes. Change the default 
\code{download.file} if the download is failing on your platform.
}
\examples{
lon <- c(-90, -60)
lat <- c(0, 30)
time <- as.Date('2013-03-01')
get.hycom(lon, lat, time, type='a', filename = '', vars = 'water_temp')
# only returns url because filename is unspecified
\dontrun{ 
get.hycom(lon, lat, time, type='a', filename = 'my_data.nc', vars = 'water_temp')
nc <- open.nc('my_data.nc')
hycom <- var.get.nc(nc, 'water_temp')
image.plot(hycom[,,1])
}
  
}
\author{
Function originally written for R by Ben Jones (WHOI) and modified
  by Camrin Braun and Ben Galuardi.
}
\references{
\url{https://hycom.org/}
}

