% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nll.hmm.r
\name{get.nll.fun}
\alias{get.nll.fun}
\title{Negative Log likelihood of parameters}
\usage{
get.nll.fun(parvec = c(D1, D2, p), g, L, dt)
}
\arguments{
\item{parvec}{vector of length 6 contianing
* kernel 1
* kernel 2
* diagonal of 2x2 matrix}

\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{final likelihood (2D)}

\item{dt}{time step}
}
\value{
parameter values
}
\description{
Negative Log likelihood of parameters
}
\examples{
# NOT RUN

 par0 <- c(log(10), log(10), log(0.5), log(0.5), log(0.95/0.05), log(0.95/0.05))
 fit <- nlm(neg.log.lik.fun, par0, L, dt)
 D1 <- exp(fit$estimate[1:2])
 D2 <- exp(fit$estimate[3:4])
 p <- 1/(1+exp(-fit$estimate[5:6])) 

}

