% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.filter.r
\name{hmm.filter}
\alias{hmm.filter}
\title{HMM filter and smoother functions}
\usage{
hmm.filter(g, L, K1, K2, P)
}
\arguments{
\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{final likelihood (2D)}

\item{K1}{first movement (diffusion) kernel see \code{\link{gausskern}}}

\item{K2}{second movement (diffusion) kernel see \code{\link{gausskern}}}

\item{P}{2x2 probability matrix for tranisitons between states (K1 and K2)}
}
\value{
a list: list(phi = phi, pred = pred, psi = psi) where
\itemize{
 \item phi is the probability for each state at each trime step 
 \item pred is ....
 \item psi is.... 
}
}
\description{
HMM filter and smoother functions
}
\examples{
P <- matrix(c(p[1], 1-p[1], 1-p[2], p[2]), 2, 2, byrow=TRUE)
}

