\name{hmmwoa-package}
\alias{hmmwoa-package}
\docType{package}
\title{
hmmwoa: A package to 
}
\description{
This package does some cool things...
}
\details{
\tabular{ll}{
Package: \tab hmmwoa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-04-26\cr
License: \tab MIT\cr
}


Long description here


}
\author{
Camrin Braun and Benjamin Galuardi

Maintainer: Who to complain to <yomama@gmail.com>

}
\references{
Pedersen, M.W., Patterson, T.A., Thygesen, U.H., and Madsen, H. 2011. Estimating animal behavior and residency from movement data. Oikos 120(9): 1281-1290. doi:10.1111/j.1600-0706.2011.19044.x.

Pedersen, M.W., Righton, D., Thygesen, U.H., Andersen, K.H., and Madsen, H. 2008. Geolocation of North Sea cod (Gadus morhua) using hidden Markov models and behavioural switching. Canadian Journal of Fisheries and Aquatic Sciences 65(11): 2167-1377. doi:10.1139/F08-144&sid=libx&genre=article.

Le Bris, A.L., Frechet, A., Galbraith, P.S., and Wroblewski, J.S. 2013. Evidence for alternative migratory behaviours in the northern Gulf of St Lawrence population of Atlantic cod (Gadus morhua L.). ICES J. Mar. Sci. doi:10.1093/icesjms/fst068.


}

\keyword{hidden markov model}
\keyword{world ocean atlas}
\keyword{ocean heat content}
\keyword{geolocation}



