% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.L.r
\name{make.L}
\alias{make.L}
\title{Combine individual source likelihoods}
\usage{
make.L(L1, L2 = NULL, L3 = NULL, known.locs = NULL, L.mle.res,
  dateVec = NULL, locs.grid = NULL, iniloc = NULL)
}
\arguments{
\item{L1}{a likelihood array}

\item{L2}{a likelihood array}

\item{L3}{a likelihood array}

\item{L.mle.res}{is a coarse resolution array of dim(L1) that speeds up the parameter estimation step later on}

\item{L.locs}{is array of dim(L1) that contains any known locations during the deployment that will supersede any other likelihood at that time point}

\item{plot}{is logical indicating whether you want an example plot}
}
\value{
a list containing: L, the overall likelihood array and L.mle, a more coarse version of L used later for parameter estimation
}
\description{
\code{make.L} combines individual likelihoods from various data sources (e.g. SST, OHC) to make overall combined likelihoods for each time point
}
\note{
This function currently only supports the use of 3 input likelihood data sources. This will be expanded in the future based on user needs.
}
\examples{
none
}

