% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshgrid.r
\name{meshgrid}
\alias{meshgrid}
\title{Creates grid from matrices}
\usage{
meshgrid(x, y)
}
\arguments{
\item{x}{vector, usually of longitude data}

\item{y}{vector, usually of latitude data}
}
\value{
list of 2 matrices for lon/lat values
}
\description{
Creates grid from matrices
}
\examples{
x <- c(1, 2, 3)
y <- c(10, 9, 8)
meshgrid(x, y)

}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
Estimating animal behavior and residency from movement data. Oikos
120:1281–1290. doi: 10.1111/j.1600-0706.2011.19044.x
}

