% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.r
\name{repmat}
\alias{repmat}
\title{Repeat your matrix?}
\usage{
repmat(X, m, n)
}
\arguments{
\item{m}{...}

\item{n}{...}

\item{x}{...}
}
\value{
matrix ...
}
\description{
Repeat your matrix?
}
\examples{
x <- c(1, 2, 3)
repmat(x)

}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
Estimating animal behavior and residency from movement data. Oikos
120:1281–1290. doi: 10.1111/j.1600-0706.2011.19044.x
}

