% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.grid.R
\name{resample.grid}
\alias{resample.grid}
\title{Resample likelihood rasters to common resolution/extent}
\usage{
resample.grid(L.rasters, res)
}
\arguments{
\item{L.rasters}{list of individual likelihood rasters generated by calc
functions}

\item{res}{raster or raster brick indicating desired output resolution
of all likelihood rasters.}
}
\value{
a list of all resampled likelihood rasters and g, the common grid
}
\description{
Resample likelihood rasters to common resolution/extent
}
\examples{
none

}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}

