% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.locs.R
\name{calc.locs}
\alias{calc.locs}
\title{Calculate Position-based Likelihood}
\usage{
calc.locs(locs, gps = NULL, iniloc, locs.grid, dateVec, errEll = TRUE,
  gpeOnly = TRUE)
}
\arguments{
\item{locs}{is -Locations file output from DAP/Tag Portal for WC tags and 
contains GPS, Argos, and GPE locations as applicable.}

\item{gps}{is -FastGPS file output from WC Tag Portal}

\item{iniloc}{is 2 x 5 dataframe containing day, month, year, lat, lon for 
both tag and pop locations}

\item{locs.grid}{is list output from \code{setup.locs.grid}}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}

\item{errEll}{is logical indicating whether error ellipses should be 
generated for light-based likelihoods as given from output of WC-GPE. False
if only longitude should be used. If False, standard deviation on light 
measurements is currently fixed at 0.7 deg longitude following Musyl et al 
2001. Default is FALSE and will use longitude only.}

\item{gpeOnly}{is logical. If TRUE, locs input is trimmed to GPE positions
only. This is most applicable in scenarios with FastGPS data and you're
adding this as a gps input (see "gps" argument).}
}
\value{
L is an array of lon x lat likelihood surfaces (matrices) for each 
  time point (3rd dimension)
}
\description{
\code{calc.locs} calculates likelihood estimates for each day of animal tag 
data.
}
\details{
Light errors are parameterized using elliptical error values output in 
'-Locations.csv' (WC tags). GPS and Argos positions are also given a 
"likelihood" using this function but are currently both considered to be 
"known" positions without error.
}

