% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nll.fun.r
\name{get.nll.fun}
\alias{get.nll.fun}
\title{Negative Log likelihood of parameters}
\usage{
get.nll.fun(parvec = c(10, 30, 5, 2, 0.707, 0.8), g = g.mle, L = L.mle)
}
\arguments{
\item{parvec}{vector of length 6 contianing * kernel 1 * kernel 2 * diagonal
of 2x2 matrix}

\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{final likelihood (2D)}
}
\value{
parameter values
}
\description{
Negative Log likelihood of parameters
}
\examples{
# NOT RUN
par0 <- c(log(10), log(10), log(0.5), log(0.5), log(0.95/0.05), log(0.95/0.05))
fit <- nlm(get.nll.fun, par0, g, L, dt)
D1 <- exp(fit$estimate[1:2])
D2 <- exp(fit$estimate[3:4])
p <- 1/(1+exp(-fit$estimate[5:6])) 
}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}

