% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.locs.grid.R
\name{setup.locs.grid}
\alias{setup.locs.grid}
\title{Setup the discrete spatial grid for the HMM}
\usage{
setup.locs.grid(limits, res = "quarter")
}
\arguments{
\item{limits}{can take either a 'data.frame' of -Locations from WC psat tag
or a pre-determined list of limits as \code{list(xmin, xmax, ymin, ymax)}}

\item{res}{character indicating resolution of grid. 'hycom' is .08 to match
hycom reanalysis res. 'quarter' and 'one' are .25 and 1 deg, respectively.}
}
\value{
a list
}
\description{
Setup the discrete spatial grid for the HMM
}
\examples{
none

}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}

